third_party/blink/web_tests/storage/indexeddb/mozilla/open-database-null-name-expected.txt

Test IndexedDB: should NOT throw when opening a database with a null name

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

indexedDB.open(null);
db = event.target.result
PASS db.name is 'null'
PASS successfullyParsed is true

TEST COMPLETE
